Rename pages to routes folder
